float gives better accuracy for 1 million iterations

+++ src/leibniz.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 
 unsigned rounds;
-double x = 1.0;
+float x = 1.0f;
 double pi = 1.0;
 
 int main()
